improve e2e tests
