+++ e2e/protractor.conf.js
@@ -40,6 +40,20 @@ exports.config = {
             }
         }));
         browser.manage().timeouts().implicitlyWait(10000);
+
+        const origFn = browser.driver.controlFlow().execute;
+
+        browser.driver.controlFlow().execute = function() {
+            const args = arguments;
+
+            // queue 100ms wait
+            origFn.call(browser.driver.controlFlow(), function() {
+                return protractor.promise.delayed(100);
+            });
+
+            return origFn.apply(browser.driver.controlFlow(), args);
+          };
+
         return browser.get('/');
     }
 };
